# api-services/swagger.yaml
swagger: "2.0"
info:
  version: "0.9"
  title: Carpooling platform API
# during dev, should point to your local machine
host: localhost:4000
...
basePath: /api/v1
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /message:
    post:
      description: Send a message
      # used as the method name of the controller
      operationId: sendMessage
      consumes:
        - application/json
      security:
        - Bearer: []
      parameters:
        - in: body
          name: message
          required: true
          description: A JSON object containing a message
          schema:
            $ref: "#/definitions/Message"
      responses:    
        "200":
              description: OK
              schema:
                description: A JSON object containing the idMessage
                example:
                  {
                    "idMessage": "jgj54qw"
                  }  
  /message/{idMessage}:
    put:
      description: Update message status
      operationId: updateMessageStatus
      security:
        - Bearer: []
      parameters:
        - in: body
          name: status
          required: true
          description: the new status of the message
          schema:
            $ref: "#/definitions/MessageStatus"
        - in: path
          name: idMessage
          required: true
          type: string
          description: the id of the message
      responses:
        "200":
          description: OK
...
  /proof:
    post:
      description: Upgrade a proof class.
      # used as the method name of the controller
      operationId: upgradeProof
      consumes:
        - application/json
      security:
        - Bearer: []
      parameters:
        - in: body
          name: proof
          required: true
          description: A JSON object containing transaction and gps.
          schema:
            $ref: "#/definitions/PostProofBody"
      responses:
        "200":
          description: OK
  /proof/{idTrip}:
     get:
      description: Get the proofs of a specific trip.
      # used as the method name of the controller
      operationId: getProofByTrip
      consumes:
        - application/json
      security:
        - Bearer: []
      parameters:
        - name: idTrip
          in: path
          description: The id of the trip.
          required: true
          type: string
      responses:
        "200":
          description: OK
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Proof"
...
  /kms:
    post:
      description: Create key pair.
      # used as the method name of the controller
      operationId: createKeys
      consumes:
        - application/json
      security:
        - Bearer: []
      parameters:
        - in: body
          name: proof
          required: true
          description: A JSON object containing idOperator.
          schema:
            $ref: "#/definitions/PostKmsKeysBody"
      responses:
        "200":
          description: OK
  /kms/rekey:
    post:
      description: Create re-encryption key.
      # used as the method name of the controller
      operationId: createRekey
      consumes:
        - application/json
      security:
        - Bearer: []
      parameters:
        - in: body
          name: proof
          required: true
          description: A JSON object containing operators delegator/delegatee and keys.
          schema:
            $ref: "#/definitions/PostKmsRekeyBody"
      responses:
        "200":
          description: OK
  /kms/{idOperator}:
    get:
      description: Get keys for an operator.
      # used as the method name of the controller
      operationId: getkeysById
      consumes:
        - application/json
      security:
        - Bearer: []
      parameters:
        - name: idOperator
          in: path
          description: The id of the operator
          required: true
          type: string
      responses:
        "200":
          description: OK
definitions:
...
  Message:
    required:
      - idOffer
      - idOperator
      - idPassenger
      - passengerShortname
      - idOperatorPassenger
      - messageTxt
      - created_at
    properties:
      idOffer:
        type: string
      idOperator:
        type: string
        enum: [op1,op2]
      idPassenger:
        type: string 
      passengerShortname:
        type: string
      idOperatorPassenger:
        type: string
        enum: [op1,op2]
      messageTxt:
        type: string
      created_at:
        type: number
  MessageStatus:
      required:
        - state
      properties:
        state:
          type: string
          enum: [seen,deleted,sent]
...
  Proof:
    properties:
      idProof:
        type: string
      idTrip:
        type: string
      idDriver:
        type: string
      driverShortname:
        type: string
      idOperatorDriver:
        type: string
        enum: [op1,op2]
      idPassenger:
        type: string
      passengerShortname:
        type: string
      idOperatorPassenger:
        type: string
        enum: [op1,op2]
      origin:
        type: string
      destination:
        type: string
      date:
        type: string
        format: date-time
  PostProofBody:
    required:
      - idOffer
      - idOperator
      - type
      - gps
      - phone

    properties:
      idOffer:
        type: string
      idOperator:
        type: string
        enum: [op2,op1]
      type:
        type: string
        enum: [TRIP_STARTED, TRIP_ENDED]
      gps:
        type: string
      phone:
        type: string
      idPassenger:
        type: string
      passengerShortname:
        type: string
      idOperatorPassenger:
        type: string
        enum: [op2,op1]
    example:
                idOffer: 23568
                idOperator: op1
                idPassenger: 56984
                passengerShortname: BBob
                idOperatorPassenger: op2
                type: TRIP_STARTED
                gps: "125, 678"
  PostKmsKeysBody:
    required:
      - idOperator

    properties:
      idOperator:
        type: string
        enum: [op2,op1]
    example:
                idOperator: op1
  PostKmsRekeyBody:
    required:
      - idOperatorDelegator
      - idOperatorDelegatee

    properties:
      idOperatorDelegator:
        type: string
        enum: [op2,op1]
      idOperatorDelegatee:
        type: string
        enum: [op1,op2]